+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.0]
+### Features
 
-* TODO: Describe initial release.
+- Added test cases for Create Checkout and Create Single Payment
+- Added **PaymayaResponse** for api response object 
+### BUG FIXES
+- Fix documentation markdown
+- Fix default values for `toMap`
+- Fix required fields
+## [0.0.1]
+- Added PayMaya SDK port from https://github.com/PayMaya/PayMaya-JS-SDK-v2